refactor(manager): rename some structs and packages and adding some questions as TODOs
